rename identity_update_transition module
